refactor: fsd 적용
