+++ src/app/model/dialog-store.ts
