+++ src/entities/comment/model/type.ts
