+++ src/entities/post/model/type.ts
