+++ src/entities/tag/model/type.ts
