+++ src/features/tags/api/useTags.ts
@@ -1,0 +1,12 @@
+import { getTags } from "../../../entities/tag/api/index"
+import { useQuery } from "@tanstack/react-query"
+
+export const useTags = () => {
+  const { data: tags, error, isLoading } = useQuery({ queryKey: ["tags"], queryFn: getTags })
+
+  return {
+    tags,
+    error,
+    loading: isLoading,
+  }
+}
