--- src/hooks/useTags.ts
@@ -1,12 +1,0 @@
-import { getTags } from "../api/tag"
-import { useQuery } from "@tanstack/react-query"
-
-export const useTags = () => {
-  const { data: tags, error, isLoading } = useQuery({ queryKey: ["tags"], queryFn: getTags })
-
-  return {
-    tags,
-    error,
-    loading: isLoading,
-  }
-}
