+++ src/widgets/post/ui/PostTable.tsx
@@ -1,0 +1,121 @@
+import { Edit2, MessageSquare, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react"
+import { Button, Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "../../../shared/ui"
+import { highlightText } from "../../../app/lib/html"
+import { getUser } from "../../../entities/user/api"
+import { PostWithUser } from "../../../entities/post/model/type"
+import { PartialUser, User } from "../../../entities/user/model/type"
+import { useDialogStore } from "../../../app/model/dialog-store"
+import { useParamsStore } from "../../../app/model/params-store"
+import { useNavigate } from "react-router-dom"
+import { usePosts } from "../../../features/posts/api/usePosts"
+
+interface Props {
+  onSelectPost: (post: PostWithUser | null) => void
+  onSelectUser: (user: User | null) => void
+}
+
+export const PostTable = ({ onSelectPost, onSelectUser }: Props) => {
+  const { posts, deletePost } = usePosts()
+
+  const { onOpenChange } = useDialogStore()
+  const { searchQuery, selectedTag, updateURL, setParams } = useParamsStore()
+  const navigate = useNavigate()
+
+  const openPostDetail = (post: PostWithUser) => {
+    onSelectPost(post)
+    onOpenChange("postDetailDialog", true)
+  }
+
+  const openUserModal = async (user: PartialUser) => {
+    try {
+      const data = await getUser(user.id) // ! 굳이....
+      onSelectUser(data)
+      onOpenChange("userModal", true)
+    } catch (error) {
+      console.error("사용자 정보 가져오기 오류:", error)
+    }
+  }
+
+  return (
+    <Table>
+      <TableHeader>
+        <TableRow>
+          <TableHead className="w-[50px]">ID</TableHead>
+          <TableHead>제목</TableHead>
+          <TableHead className="w-[150px]">작성자</TableHead>
+          <TableHead className="w-[150px]">반응</TableHead>
+          <TableHead className="w-[150px]">작업</TableHead>
+        </TableRow>
+      </TableHeader>
+      <TableBody>
+        {posts &&
+          posts.posts.map((post) => (
+            <TableRow key={post.id}>
+              <TableCell>{post.id}</TableCell>
+              <TableCell>
+                <div className="space-y-1">
+                  <div>{highlightText(searchQuery, post.title)}</div>
+
+                  <div className="flex flex-wrap gap-1">
+                    {post.tags?.map((tag) => (
+                      <span
+                        key={tag}
+                        className={`px-1 text-[9px] font-semibold rounded-[4px] cursor-pointer ${
+                          selectedTag === tag
+                            ? "text-white bg-blue-500 hover:bg-blue-600"
+                            : "text-blue-800 bg-blue-100 hover:bg-blue-200"
+                        }`}
+                        onClick={() => {
+                          setParams("selectedTag", tag)
+                          updateURL(navigate)
+                        }}
+                      >
+                        {tag}
+                      </span>
+                    ))}
+                  </div>
+                </div>
+              </TableCell>
+              <TableCell>
+                <div
+                  className="flex items-center space-x-2 cursor-pointer"
+                  onClick={() => post.author && openUserModal(post.author)}
+                >
+                  <img src={post.author?.image} alt={post.author?.username} className="w-8 h-8 rounded-full" />
+                  <span>{post.author?.username}</span>
+                </div>
+              </TableCell>
+              <TableCell>
+                <div className="flex items-center gap-2">
+                  <ThumbsUp className="w-4 h-4" />
+                  <span>{post.reactions?.likes || 0}</span>
+                  <ThumbsDown className="w-4 h-4" />
+                  <span>{post.reactions?.dislikes || 0}</span>
+                </div>
+              </TableCell>
+              <TableCell>
+                <div className="flex items-center gap-2">
+                  <Button variant="ghost" size="sm" onClick={() => openPostDetail(post)}>
+                    <MessageSquare className="w-4 h-4" />
+                  </Button>
+                  <Button
+                    variant="ghost"
+                    size="sm"
+                    onClick={() => {
+                      onSelectPost(post)
+                      onOpenChange("editPostDialog", true)
+                    }}
+                  >
+                    <Edit2 className="w-4 h-4" />
+                  </Button>
+                  <Button variant="ghost" size="sm" onClick={() => deletePost(post.id)}>
+                    <Trash2 className="w-4 h-4" />
+                  </Button>
+                </div>
+              </TableCell>
+            </TableRow>
+          ))}
+      </TableBody>
+    </Table>
+  )
+}
